+++ ShinRyuModManager/LibMeta.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Net;
+using System.IO;
 using YamlDotNet.Serialization;
 
 namespace ShinRyuModManager
@@ -44,6 +46,23 @@ namespace ShinRyuModManager
             return meta;
         }
 
+
+        public static List<LibMeta> Fetch()
+        {
+            WebClient client = new WebClient();
+            //TODO UNCOMMENT THIS WHEN THE REPO IS READY. READING FROM LOCAL FILE IN THE MEANTIME
+            string yamlString = client.DownloadString($"https://raw.githubusercontent.com/{Settings.LIBRARIES_INFO_REPO_OWNER}/{Settings.LIBRARIES_INFO_REPO}/main/{Settings.LIBRARIES_INFO_REPO_FILE_PATH}");
+            //string yamlString = File.ReadAllText(Program.GetLocalLibraryCopyPath());
+
+            // Save a copy of the manifest for offline use
+            string localManifestCopyPath = Program.GetLocalLibraryCopyPath();
+            if (!File.Exists(localManifestCopyPath) && !Util.IsFileBlocked(localManifestCopyPath))
+            {
+                File.WriteAllText(localManifestCopyPath, yamlString);
+            }
+
+            return LibMeta.ReadLibMetaManifest(yamlString);
+        }
 
         /// <summary>
         /// Reads the <see cref="LibMeta"/> manifest from a YAML string.